Refactor for configuration file
Abstracted the Excel spreadsheet mapping to a JSON configuration file that is read at runtime.

Created a JSON schema and added validation of the configuration file against the schema before attempting to generate HTML output files.

Improved the Jinja templates to correct accessibility issues and present the values better.

Added templates and configuration for the Higher Education Emergency Relief Fund.